pdfsave/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cleanup==5.1.0
lxml==4.6.2
pkg-resources==0.0.0
PyMySQL==1.0.2
python-docx==0.8.10
pytz==2020.5
sqlparse==0.4.1
urllib3==1.26.2
